refactor(server): move the argument validators into src/server/tool-validation.ts
The sessionId guard was copied into fourteen handlers in two message variants,
the "must be an object when provided" check into three, and
validateBreakOnExceptions / normalizeStartDebuggingArgs / enforceExplicitNames
sat on DebugMcpServer purely so the handlers could reach them through the
context — none of them touch session state.

tool-validation.ts now owns requireSessionId (an assertion function, so the
handlers keep their narrowed args), assertPlainObjectArg, the two start_debugging
normalizers with NEVER_VALID_DAP_LAUNCH_KEYS, and enforceExplicitNames over an
explicit IEnvironment. They leave DebugMcpServer and ToolContext.

One deliberate user-visible change: the two single-parameter sessionId messages
('Missing required sessionId' and 'Missing required parameter: sessionId') are
unified to 'Missing required parameter: sessionId'. The multi-parameter guards
keep 'Missing required parameters'. server-control-tools.test.ts pinned the old
wording and is updated.
