+++ src/server.ts
@@ -48,7 +48,7 @@ import {
   supportsLoudSnapping
 } from './utils/bp-addressing.js';
 import { isRedactionEnabled } from './utils/redaction-mode.js';
-import { getVariableAccessMode, requiresExplicitNames } from './utils/variable-access.js';
+import { getVariableAccessMode } from './utils/variable-access.js';
 import { assertLineContent, resolveStatement, stripTrailingComment } from './utils/breakpoint-resolver.js';
 import { OutputResourceNotifier, registerResourceHandlers } from './server/output-resources.js';
 import { registerPromptHandlers } from './server/prompts.js';
@@ -224,91 +224,6 @@ export class DebugMcpServer implements ToolContext {
     } catch {
       return undefined;
     }
-  }
-
-  /** @internal ToolContext service. */
-  public validateBreakOnExceptions(value: string | undefined): ExceptionBreakMode | undefined {
-    if (value === undefined) {
-      return undefined;
-    }
-    if (value !== 'uncaught' && value !== 'all' && value !== 'none') {
-      throw new McpError(
-        McpErrorCode.InvalidParams,
-        `breakOnExceptions must be one of 'uncaught', 'all', 'none' (got '${value}')`
-      );
-    }
-    return value;
-  }
-
-  /**
-   * Top-level start_debugging parameters that have no meaning inside
-   * dapLaunchArgs. Deliberately excludes keys that are legitimate DAP launch
-   * arguments (program, cwd, args, env, stopOnEntry, justMyCode, ...) —
-   * compiled languages pass the binary as dapLaunchArgs.program.
-   * breakOnExceptions is handled separately (honored as an alias).
-   */
-  private static readonly NEVER_VALID_DAP_LAUNCH_KEYS = [
-    'dryRunSpawn',
-    'sessionId',
-    'scriptPath',
-    'adapterLaunchConfig',
-    'dapLaunchArgs'
-  ] as const;
-
-  /**
-   * Intake normalization for start_debugging (issue #305). dapLaunchArgs is
-   * declared additionalProperties:true, so a top-level parameter nested there
-   * by mistake used to ride through the launch-config merge as a junk DAP key
-   * adapters silently ignore — a silent behavioral failure. Now:
-   * - dapLaunchArgs.breakOnExceptions is honored as an alias for the
-   *   top-level parameter (top-level wins when both are given), stripped from
-   *   the forwarded launch args, and reported via a warning.
-   * - Other never-valid nested keys are stripped with a warning.
-   * Fixing at intake also cures restart_debugging replay, which snapshots the
-   * post-intake values into session.lastLaunch downstream.
-   * @internal ToolContext service.
-   */
-  public normalizeStartDebuggingArgs(
-    dapLaunchArgs: Partial<DebugProtocol.LaunchRequestArguments> | undefined,
-    topLevelBreakOnExceptions: string | undefined
-  ): {
-    dapLaunchArgs: Partial<DebugProtocol.LaunchRequestArguments> | undefined;
-    breakOnExceptions: string | undefined;
-    warnings: string[];
-  } {
-    const warnings: string[] = [];
-    let breakOnExceptions = topLevelBreakOnExceptions;
-    if (dapLaunchArgs === null || typeof dapLaunchArgs !== 'object' || Array.isArray(dapLaunchArgs)) {
-      return { dapLaunchArgs, breakOnExceptions, warnings };
-    }
-    const cleaned: Record<string, unknown> = { ...dapLaunchArgs };
-    if ('breakOnExceptions' in cleaned) {
-      const nested = cleaned.breakOnExceptions;
-      delete cleaned.breakOnExceptions;
-      if (topLevelBreakOnExceptions === undefined) {
-        breakOnExceptions = nested as string;
-        warnings.push(
-          `breakOnExceptions is a top-level start_debugging parameter, not a dapLaunchArgs key — honored as '${String(nested)}' this time; pass it at the top level`
-        );
-      } else {
-        warnings.push(
-          `breakOnExceptions was passed both top-level ('${topLevelBreakOnExceptions}') and inside dapLaunchArgs ('${String(nested)}'); the top-level value wins and the nested key was ignored`
-        );
-      }
-    }
-    for (const key of DebugMcpServer.NEVER_VALID_DAP_LAUNCH_KEYS) {
-      if (key in cleaned) {
-        delete cleaned[key];
-        warnings.push(
-          `'${key}' is a top-level start_debugging parameter and has no meaning inside dapLaunchArgs — ignored; pass it at the top level`
-        );
-      }
-    }
-    return {
-      dapLaunchArgs: cleaned as Partial<DebugProtocol.LaunchRequestArguments>,
-      breakOnExceptions,
-      warnings
-    };
   }
 
   // Core business logic methods — delegated to by the MCP tool dispatch handler
@@ -801,26 +716,6 @@ export class DebugMcpServer implements ToolContext {
   /** @internal test seam; removed in PR 6 */
   private async handleListDebugSessions(): Promise<ToolResult> {
     return handleListDebugSessions(this);
-  }
-
-  /**
-   * Least-privilege enforcement (issue #237): in explicit mode, bulk scope
-   * dumps are disabled — the tools require a non-empty names filter, and the
-   * error teaches the correct call shape.
-   * @internal ToolContext service.
-   */
-  public enforceExplicitNames(toolName: string, names: string[] | undefined): void {
-    if (!requiresExplicitNames(getVariableAccessMode(this.environment))) {
-      return;
-    }
-    if (!names || names.length === 0) {
-      throw new McpError(
-        McpErrorCode.InvalidParams,
-        `${toolName} requires "names" in least-privilege mode (DEBUG_MCP_VARIABLE_ACCESS=explicit): ` +
-        `pass the exact variable names you need, e.g. names:["user","order_total"]. ` +
-        `Unfiltered scope dumps are disabled by this server's configuration.`
-      );
-    }
   }
 
   /**
